feat: implemented listener for shiny stat changes.
fix: shiny mythic stats changing in game now modify the shiny mythic stat stored in the gear instance

+++ common/src/main/java/com/wynntils/models/wynnitem/parsing/WynnItemParseResult.java
@@ -24,4 +24,4 @@ public record WynnItemParseResult(
         int rerolls,
         int durabilityCurrent,
         int durabilityMax,
-        Optional<Pair<String, Integer>> shinyStat) {}
+        Optional<Pair<String, Long>> shinyStat) {}
